[jax2tf] Handle weakly-typed values in constant lifting correctly.
An upcoming change to JAX will add `jit` decorators around many standard library functions. This was causing the pattern in this test case to break: the test before this change returns a float64 value. However, no type promotion should take place because 2. is a Python scalar and should be weakly-typed.